Revert "Merge branch 'master' of https://github.com/nebw/drunken-markov"
This reverts commit e50abe8d6f412060301c58c47e1fd66a85a650c8, reversing
changes made to d2325283a2332a7d2149bd3557d3ba70f69d8b52.

+++ drunkenmarkov/tests/TPT_test.py
@@ -16,7 +16,7 @@ class TPTTests(TestCase):
         in early 2015 at FU Berlin.
         """
 
-        rate_matrix = np.array([[-1, 1, 0], [100, -300, 200], [0, 1, -1]]).astype(float)
+        rate_matrix = np.array([[-1, 1, 0], [100, -300, 200], [0, 1, -1]])
         K = scipy.linalg.expm(rate_matrix)
         TPT = TransitionPathTheory(K, [0], [1])
 
@@ -32,7 +32,7 @@ class TPTTests(TestCase):
         in early 2015 at FU Berlin.
         """
 
-        rate_matrix = np.array([[-1, 1, 0], [100, -300, 200], [0, 1, -1]]).astype(float)
+        rate_matrix = np.array([[-1, 1, 0], [100, -300, 200], [0, 1, -1]])
         K = scipy.linalg.expm(rate_matrix)
         TPT = TransitionPathTheory(K, [0], [1])
